% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_1Mn.R
\name{SE_1Mn}
\alias{SE_1Mn}
\title{creates dense SummarizedExperiment with the 1Mneuron data}
\usage{
SE_1Mn(samps = 1:100,
  h5fname = "./1M_neurons_filtered_gene_bc_matrices_h5.h5")
}
\arguments{
\item{samps}{a vector of sample indices to use (might be able to use barcodes)}

\item{h5fname}{path to hdf5 for the 10x genomics data as distributed}
}
\author{
VJ Carey
}
