% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxGenomics-class.R
\docType{methods}
\name{TENxGenomics}
\alias{TENxGenomics}
\alias{show,TENxGenomics-method}
\title{Create and manipulate a reference to 10xGenomics data.}
\usage{
TENxGenomics(h5path)

\S4method{show}{TENxGenomics}(object)
}
\arguments{
\item{h5path}{character(1) file path to the
1M_neurons_filtered_gene_bc_matrices_h5.h5 file.}

\item{object}{A \code{TENxGenomics-class} instance.}
}
\value{
\code{TENxGenomics()} returns a \code{TENxGenomics-class}
    instance.
}
\description{
The TENxGenomics class provides a simple interface to
    subset and input 10xGenomis'
    1M_neurons_filtered_gene_bc_matrices_h5.h5 file. Subsetting is
    a light-weight operation; input (typically of the subset
    matrix) is as a dense matrix and hence consumes memory.
}
