% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenxiterate.R
\name{tenxiterate}
\alias{tenxiterate}
\title{Visit and 'reduce()' all values in a 10xGenomics hdf5 file.}
\usage{
tenxiterate(X, FUN, ..., yieldSize = 10000)
}
\arguments{
\item{X}{A \code{TENxGenomics-class} instance. Rows and columns of
\code{X} cannot be duplicated.}

\item{FUN}{A function of at least one argument \code{VALUE}, used
to summarize successive chunks of the hdf5 file. \code{VALUE}
is a list with elements \code{ridx} (row index), \code{cidx}
(column index) and \code{value} (count value). See
[BiocParallel::bpiterate()].}

\item{...}{Additional arguments, passed to \code{FUN()}.}

\item{yieldSize}{numeric(1) Number of columns (cells) to process
during each iteration.}
}
\description{
This function visits all records in the hdf5 file,
    passing a 'chunk' containing a list of value, row index, and
    column index to a user-provided \code{FUN()}. The chunks
    represent one or more complete cells (columns); the number of
    cells is controled by \code{yieldSize}. \code{FUN()}
    manipulates the arguments (typically a data reduction) in an
    arbitary way, returning the result.

    All records in the hdf5 file are visited, but only those
    correspond to rows or columns of \code{X} are reported. The
    implementation uses [BiocParallel::bpiterate()]
}
