% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenxSummarizedExperiment.R
\name{tenxSummarizedExperiment}
\alias{tenxSummarizedExperiment}
\title{Create a 'SummarizedExperiment' from an 10xGenomics hdf5 file.}
\usage{
tenxSummarizedExperiment(h5path, rowData, colData)
}
\arguments{
\item{h5path}{character(1) file path to a 1M_neurons_*.h5 file.}

\item{rowData}{A \code{DataFrame()} with as many rows as there are
genes in the 10xGenomics file. If missing, an object is created
with 'gene' and 'genename' fields from the hdf5 file.}

\item{colData}{A \code{DataFrame()} with as many rows as there are
samples in the 10xGenomics file. If missing, and object is
constructed from the barcodes of the hdf5 file. The sequence
and library portions are separated, and the mouse is inferred
(libraries <= 69 are from mouse "A", others are from mouse
"B").}
}
\value{
A \code{SummarizedExperiment} object. For effective use of
    the return object, load the SummarizedExperiment library into
    the R session.
}
\description{
The SummarizedExperiment \code{assay()} contains the
    \code{TENxGenomics} object corresponding to the underlying hdf5
    file. It also contains row and column annotations provided by
    the user or inferred from the hdf5 file. Inferred data requires
    a simple match between the file name prefix and
    \dQuote{1M_neurons_}; if the file name prefix does not match,
    row and column annotations are not created.
}
