% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxGenomics-class.R
\docType{methods}
\name{TENxGenomics}
\alias{TENxGenomics}
\alias{dim,TENxGenomics-method}
\alias{dimnames,TENxGenomics-method}
\alias{[,TENxGenomics-method}
\alias{[,TENxGenomics,ANY,ANY-method}
\alias{show,TENxGenomics-method}
\title{Create and manipulate a reference to 10xGenomics data.}
\usage{
TENxGenomics(h5path)

\S4method{dim}{TENxGenomics}(x)

\S4method{dimnames}{TENxGenomics}(x)

\S4method{[}{TENxGenomics}(x, i, j, ..., drop = TRUE)

\S4method{show}{TENxGenomics}(object)
}
\arguments{
\item{h5path}{character(1) file path to the
1M_neurons_filtered_gene_bc_matrices_h5.h5 file.}

\item{x}{A \code{TENxGenomics-class} instance.}

\item{i}{integer(), numeric(), or character() index into rows of
\code{x}.}

\item{j}{integer(), numeric(), or character() index into columns of
\code{x}.}

\item{...}{Additional arguments, ignored.}

\item{drop}{logical(1) TRUE only.}

\item{object}{A \code{TENxGenomics-class} instance.}
}
\value{
\code{TENxGenomics()} returns a \code{TENxGenomics-class}
    instance.
}
\description{
The TENxGenomics class provides a simple interface to
    subset and input 10xGenomis'
    1M_neurons_filtered_gene_bc_matrices_h5.h5 file. Subsetting is
    a light-weight operation; input (typically of the subset
    matrix) is as a dense matrix and hence consumes memory.
}
