% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxGenomics-class.R, R/coerce.R
\docType{methods}
\name{TENxGenomics}
\alias{TENxGenomics}
\alias{dim,TENxGenomics-method}
\alias{dimnames,TENxGenomics-method}
\alias{[,TENxGenomics-method}
\alias{[,TENxGenomics,ANY,ANY-method}
\alias{show,TENxGenomics-method}
\alias{as.matrix.TENxGenomics}
\alias{coerce,TENxGenomics,matrix-method}
\alias{as.dgCMatrix}
\alias{coerce,TENxGenomics,dgCMatrix-method}
\title{Create and manipulate a reference to 10xGenomics data.}
\usage{
TENxGenomics(h5path)

\S4method{dim}{TENxGenomics}(x)

\S4method{dimnames}{TENxGenomics}(x)

\S4method{[}{TENxGenomics}(x, i, j, ..., drop = TRUE)

\S4method{show}{TENxGenomics}(object)

\method{as.matrix}{TENxGenomics}(x, ..., withDimnames = TRUE)

as.dgCMatrix(x, ..., withDimnames = TRUE)
}
\arguments{
\item{h5path}{character(1) file path to a 1M_neurons_*.h5 file.}

\item{x}{A \code{TENxGenomics-class} instance.}

\item{i}{integer(), numeric(), character(), or logical() index into
rows of \code{x}.}

\item{j}{integer(), numeric(), character(), or logical() index into
columns of \code{x}.}

\item{...}{Additional arguments, ignored.}

\item{drop}{logical(1) TRUE only.}

\item{object}{A \code{TENxGenomics-class} instance.}

\item{withDimnames}{logical(1) Include dimnames on returned matrix?}
}
\value{
\code{TENxGenomics()} returns a \code{TENxGenomics-class}
    instance.

\code{as.matrix(tenx)} and \code{as(tenx, "matrix")} return
    a matrix with dim and dimnames equal to \code{tenx}, and values
    the read counts overlapping corresponding genes and
    samples. Use \code{as.matrix(withDimnames=FALSE)} to suppress
    dimnames on the returned matrix. NOTE: consider the size of the
    matrix, \code{prod(as.numeric(dim(tenx)))} before invoking this
    function.

\code{as.dgCMatrix(tenx)} and \code{as(tenx, "dgCMatrix")}
    return a sparse matrix (from the Matrix package) with dim and
    dimnames equal to \code{tenx}, and values the read counts
    overlapping corresponding genes and samples. Use
    \code{as.matrix(withDimnames=FALSE)} to suppress dimnames on
    the returned matrix.
}
\description{
The TENxGenomics class provides a simple interface to
    subset and input 10xGenomis'
    1M_neurons_filtered_gene_bc_matrices_h5.h5 file. Subsetting is
    a light-weight operation; input (typically of the subset
    matrix) is as a dense matrix and hence consumes memory.
}
