% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxMatrix-class.R
\docType{class}
\name{TENxMatrix-class}
\alias{TENxMatrixSeed-class}
\alias{TENxMatrix-class}
\alias{TENxMatrixSeed}
\alias{TENxMatrix}
\alias{dim,TENxMatrixSeed-method}
\alias{dimnames,TENxMatrixSeed-method}
\alias{DelayedArray,TENxMatrixSeed-method}
\title{TENxMatrix objects}
\description{
A container for representing 10xGenomics data.
    TENxMatrix extends \link{DelayedArray} so all the operations
    available on \link{DelayedArray} objects work on TENxMatrix objects.
}

